Correcciones: Borrado de clases y atributos no necesarios

+++ ERP-D/ERP-D/Models/CentroDeCosto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public double MontoMaximo { get; set; }
+        public decimal MontoMaximo { get; set; }
         public List<Gasto> Gastos { get; set; }
     }
 }
